requirements.txt
Django==1.10.3
django-anymail==0.6.1
django-grappelli==2.9.1
mysqlclient==1.3.9
Pillow==3.4.2
requests==2.12.4
six==1.10.0
